ollama-chat/requirements.txt
pyyaml
ollama
gradio
